question2/requirements.txt
certifi==2017.11.5
chardet==3.0.4
Django==1.11
djangorestframework==3.6.2
idna==2.6
minance-sms==0.1.1
pytz==2018.4
requests==2.18.4
urllib3==1.22
